+++ device_support/f28004x/headers/include/f28004x_device.h
@@ -8,7 +8,7 @@
 //
 //
 // $Copyright:
-// Copyright (C) 2025 Texas Instruments Incorporated - http://www.ti.com/
+// Copyright (C) 2026 Texas Instruments Incorporated - http://www.ti.com/
 //
 // Redistribution and use in source and binary forms, with or without 
 // modification, are permitted provided that the following conditions 
@@ -46,7 +46,7 @@
 #ifdef __cplusplus
 extern "C" {
 #endif
-
+#include "stdint.h"
 //
 // Common CPU Definitions:
 //
@@ -55,8 +55,8 @@ extern "C" {
 #define __cregister
 #endif  //__TMS320C28xx__
 
-extern __cregister volatile unsigned int IFR;
-extern __cregister volatile unsigned int IER;
+extern __cregister volatile uint16_t IFR;
+extern __cregister volatile uint16_t IER;
 
 #define  EINT   __asm(" clrc INTM")
 #define  DINT   __asm(" setc INTM")
@@ -180,17 +180,17 @@ typedef _Bool status_t;
 typedef short                                   int16;
 typedef long                                    int32;
 typedef unsigned char                           Uint8;
-typedef unsigned short                          Uint16;
+typedef uint16_t                                Uint16;
 typedef unsigned long                           Uint32;
 typedef float                                   float32;
 typedef long double                             float64;
 typedef struct { Uint32 low32; Uint32 high32; } Uint64;
 typedef struct { int32  low32; int32  high32; } int64;
 #else // __TMS320C28XX__
-typedef int                                     int16;
+typedef int16_t                                     int16;
 typedef long                                    int32;
 typedef long long                               int64;
-typedef unsigned int                            Uint16;
+typedef unsigned short                            Uint16;
 typedef unsigned long                           Uint32;
 typedef unsigned long long                      Uint64;
 typedef float                                   float32;
