+++ device_support/f28e12x/headers/include/f28e12x_device.h
@@ -8,7 +8,7 @@
 //
 //
 // $Copyright:
-// Copyright (C) 2025 Texas Instruments Incorporated - http://www.ti.com/
+// Copyright (C) 2026 Texas Instruments Incorporated - http://www.ti.com/
 //
 // Redistribution and use in source and binary forms, with or without 
 // modification, are permitted provided that the following conditions 
@@ -51,9 +51,9 @@ extern "C" {
 #ifndef __TMS320C28XX__
 #define __cregister
 #endif  //__TMS320C28xx__
-
-extern __cregister volatile unsigned int IFR;
-extern __cregister volatile unsigned int IER;
+#include "stdint.h"
+extern __cregister volatile uint16_t IFR;
+extern __cregister volatile uint16_t IER;
 
 #define  EINT   __asm(" clrc INTM")
 #define  DINT   __asm(" setc INTM")
@@ -176,17 +176,17 @@ typedef _Bool status_t;
 typedef short                                   int16;
 typedef long                                    int32;
 typedef unsigned char                           Uint8;
-typedef unsigned short                          Uint16;
+typedef uint16_t                                Uint16;
 typedef unsigned long                           Uint32;
 typedef float                                   float32;
 typedef long double                             float64;
 typedef struct { Uint32 low32; Uint32 high32; } Uint64;
 typedef struct { int32  low32; int32  high32; } int64;
 #else // __TMS320C28XX__
-typedef int                                     int16;
+typedef int16_t                                     int16;
 typedef long                                    int32;
 typedef long long                               int64;
-typedef unsigned int                            Uint16;
+typedef unsigned short                            Uint16;
 typedef unsigned long                           Uint32;
 typedef unsigned long long                      Uint64;
 #ifndef C2000_IEEE754_TYPES
