+++ device_support/f28p55x/headers/include/f28p55x_device.h
@@ -8,7 +8,7 @@
 //
 //
 // 
-// C2000Ware v6.00.01.00
+// C2000Ware v26.00.00.00
 //
 // Copyright (C) 2024 Texas Instruments Incorporated - http://www.ti.com
 //
@@ -54,8 +54,8 @@ extern "C" {
 #define __cregister
 #endif  //__TMS320C28xx__
 
-extern __cregister volatile unsigned int IFR;
-extern __cregister volatile unsigned int IER;
+extern __cregister volatile uint16_t IFR;
+extern __cregister volatile uint16_t IER;
 
 #define  EINT   __asm(" clrc INTM")
 #define  DINT   __asm(" setc INTM")
@@ -178,17 +178,17 @@ typedef _Bool status_t;
 typedef short                                   int16;
 typedef long                                    int32;
 typedef unsigned char                           Uint8;
-typedef unsigned short                          Uint16;
+typedef uint16_t                                Uint16;
 typedef unsigned long                           Uint32;
 typedef float                                   float32;
 typedef long double                             float64;
 typedef struct { Uint32 low32; Uint32 high32; } Uint64;
 typedef struct { int32  low32; int32  high32; } int64;
 #else // __TMS320C28XX__
-typedef int                                     int16;
+typedef int16_t                                     int16;
 typedef long                                    int32;
 typedef long long                               int64;
-typedef unsigned int                            Uint16;
+typedef unsigned short                            Uint16;
 typedef unsigned long                           Uint32;
 typedef unsigned long long                      Uint64;
 #ifndef C2000_IEEE754_TYPES
