+++ driverlib/f2838x/driverlib/inc/hw_types.h
@@ -6,7 +6,7 @@
 //
 //###########################################################################
 // 
-// C2000Ware v6.00.01.00
+// C2000Ware v26.00.00.00
 //
 // Copyright (C) 2024 Texas Instruments Incorporated - http://www.ti.com
 //
@@ -71,7 +71,7 @@ typedef int16_t int8_t;
 #define HWREG_BP(x)                                                           \
         __byte_peripheral_32((uint32_t *)(x))
 #define HWREGB(x)                                                             \
-        __byte((int16_t *)(x),0)
+        __byte((int *)(x),0)
 
 //
 // Emulated Bitbanded write
