+++ driverlib/f2838x/driverlib/pmbus.c
@@ -6,7 +6,7 @@
 //
 //#############################################################################
 // 
-// C2000Ware v6.00.01.00
+// C2000Ware v26.00.00.00
 //
 // Copyright (C) 2024 Texas Instruments Incorporated - http://www.ti.com
 //
@@ -518,7 +518,7 @@ PMBus_verifyPEC(uint32_t base, uint16_t *buffer, const uint16_t *crcTable,
         // hence parity is always even (0) and incremented by 2
         //
 
-        tableIndex = accumulator ^ (uint16_t)__byte((int16_t *)buffer, parity);
+        tableIndex = accumulator ^ (uint16_t)__byte((int *)buffer, parity);
         accumulator = crcTable[tableIndex];
         parity += 2U;
     }
